+++ src/components/App/App.css
@@ -1,38 +1,13 @@
 .App {
-  text-align: center;
+  background-color: #00B2FF;
+  display: grid;
+  grid-template-areas: "header nav nav"
+                        "sidebar main main"
+                        "sidebar main main";
+  grid-template-rows: 1fr 1fr 1fr;
+  height: 100vh;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.welcome {
+  grid-area: 'main';
 }
