+++ src/components/App/App.js
@@ -1,24 +1,21 @@
 import React from 'react';
-import logo from '../../logo.svg';
 import './App.css';
+import { Route, Switch } from 'react-router-dom';
+import Welcome from '../Welcome/Welcome'
+
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <Header />
+      <Sidebar />
+      <Nav />
+      <Switch>
+        <Route
+          path="/" exact
+          component={() => <Welcome className="welcome"/>}
+        />
+      </Switch>
     </div>
   );
 }
